refactor(bridge): remove mechanical analyzer findings

+++ bridge/Commands/ClickCommands.cs
@@ -683,8 +683,8 @@ public static class ClickCommands
         {
             var rect = element.BoundingRectangle;
             center = new Point(
-                (int)(rect.X + rect.Width / 2),
-                (int)(rect.Y + rect.Height / 2));
+                rect.X + rect.Width / 2,
+                rect.Y + rect.Height / 2);
         }
 
         if (JsonRpcHandler.Stealth)
