+++ bridge/Commands/SelectionCommands.cs
@@ -59,8 +59,8 @@ public static class SelectionCommands
                 // Fallback: Ctrl+Click
                 var rect = child.BoundingRectangle;
                 var center = new Point(
-                    (int)(rect.X + rect.Width / 2),
-                    (int)(rect.Y + rect.Height / 2));
+                    rect.X + rect.Width / 2,
+                    rect.Y + rect.Height / 2);
 
                 var useControl = mode == "add" || !isFirst;
                 if (useControl)
